// schemas/admin.ts
export const userRoleSchema = z.object({
  userRoleId: z
    .number()
    .min(0, { message: "User role ID must be 0 or greater" }),
  userRoleCode: z
    .string()
    .min(5, { message: "User role code must be at least 5 characters" })
    .max(50, { message: "User role code must be less than 50 characters" }),
  userRoleName: z
    .string()
    .min(5, { message: "User role name must be at least 5 characters" })
    .max(100, { message: "User role name must be less than 100 characters" }),
  remarks: z
    .string()
    .max(255, { message: "Remarks cannot exceed 255 characters" })
    .optional(),
  isActive: z.boolean(),
})

export type UserRoleSchemaType = z.infer<typeof userRoleSchema>